Atualização de models

+++ src/main/java/com/easygest/springboot/model/Compra.java
@@ -1,11 +1,13 @@
 package com.easygest.springboot.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 
@@ -24,11 +26,17 @@ public class Compra {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long codCompra;
+
     private Integer pedido;
+
     private String item;
+
     private LocalDate dataCompra;
+
     private LocalDate dataRecebimento;
-    private Float valor;
+
+    private BigDecimal valor;
+
     private Integer quantidade;
 
     @ManyToOne
